finish Random Generate Data Function !!!
YA~~~!

+++ Students_BMI_and_Scores_Manager/Form3.cs
@@ -34,13 +34,13 @@ namespace Students_BMI_and_Scores_Manager
                 f1.minCHscore = int.Parse(minCHscoreTextBox.Text);
                 f1.minENscore = int.Parse(minENscoreTextBox.Text);
                 f1.minMATHscore = int.Parse(minMATHscoreTextBox.Text);
-                f1.minHeight = double.Parse(minHeightTextBox.Text);
-                f1.minWeight = double.Parse(minWeightTextBox.Text);
+                f1.minHeight = int.Parse(minHeightTextBox.Text);
+                f1.minWeight = int.Parse(minWeightTextBox.Text);
                 f1.maxCHscore = int.Parse(maxCHscoreTextBox.Text);
                 f1.maxENscore = int.Parse(maxENscoreTextBox.Text);
                 f1.maxMATHscore = int.Parse(maxMATHscoreTextBox.Text);
-                f1.maxHeight = double.Parse(maxHeightTextBox.Text);
-                f1.maxWeight = double.Parse(maxWeightTextBox.Text);
+                f1.maxHeight = int.Parse(maxHeightTextBox.Text);
+                f1.maxWeight = int.Parse(maxWeightTextBox.Text);
                 f1.randomBigDataPageReturnStatus = 1;//各項參數成功回傳
             }
             catch
